requirements.txt
streamlit==1.38.0
ollama==0.3.3
pyyaml==6.0.2
requests==2.32.0